requirements.txt
amqp==1.4.9
anyjson==0.3.3
argh==0.26.1
billiard==3.3.0.20
cabby==0.1.4
celery==3.1.18
certifi==2019.3.9
chardet==3.0.4
colorlog==2.6.0
cybox==2.1.0.9
Django==1.11.20
django-extensions==2.1.6
django-filter==0.9.2
django-simple-menu==1.0.9
djangorestframework==3.9.2
geoip2==2.1.0
ghp-import==0.4.1
gnureadline==6.3.3
httplib2==0.9
idna==2.8
ipaddr==2.1.11
ipython==3.1.0
Jinja2==2.10.1
kombu==3.0.34
libtaxii==1.1.106
lxml==3.4.1
Markdown==2.4.1
MarkupSafe==1.1.1
maxminddb==1.2.0
mkdocs==0.11.1
omnijson==0.1.2
pathtools==0.1.2
pydotplus==2.0.2
pyparsing==2.2.0
python-dateutil==1.5
pytz==2014.10
PyYAML==5.1
requests==2.21.0
six==1.12.0
stix==1.1.1.3
typing==3.6.6
urllib3==1.24.1
watchdog==0.8.3
